Bug #126995: 1. Fill SFN/SLOT for 5G trace msg 2. Use same data structure for SFN/SLOT in 4G and 5G

+++ common/utils/LOG/log.h
@@ -292,7 +292,7 @@ typedef struct {
   uint64_t                dump_mask;
 #if 1
   uint16_t                sfn;
-  uint8_t                 sf;
+  uint32_t                sf;
 #endif
 } log_t;
